+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [Paging Extension] Switch to AndroidX Paging (#5910 by @jeffdgr8)
 
 ### Fixed
+- [IntelliJ Plugin] Fix IDE freeze caused by blocking file type detection on the EDT during VFS refresh events.
 - [SQLite Dialect] Fix Sqlite 3.38 compilation error when using Json path operators (#6070 by @griffio)
 - [SQLite Dialect] Use String type for group_concat function when using custom column type (#6082 by @griffio)
 - [Gradle Plugin] Improve performance of `VerifyMigrationTask` to stop it from hanging on complex schemas (#6073 by @Lightwood13)